+++ src/game/tests.rs
@@ -555,6 +555,7 @@ mod vintage_cube_tamiyo;
 mod vintage_cube_tamiyo_inquisitive_student;
 mod vintage_cube_tarfire;
 mod vintage_cube_tear_asunder;
+mod vintage_cube_teferi_hero_of_dominaria;
 mod vintage_cube_teferi_time_raveler;
 mod vintage_cube_tendrils;
 mod vintage_cube_territorial_kavu;